api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
channels
channels-redis